# implementation/at.electro.shop.bff/src/main/java/at/electro/shop/bff/api/notification/models/ProductMarkingRequest.java
package at.electro.shop.bff.api.notification.models;

import java.math.BigDecimal;

public class ProductMarkingRequest {
    private String user;
    private String emailAddress;
    private String product;
    private BigDecimal price;

    public String getUser() {
        return user;
    }

    public void setUser(String user) {
        this.user = user;
    }

    public String getEmailAddress() {
        return emailAddress;
    }

    public void setEmailAddress(String emailAddress) {
        this.emailAddress = emailAddress;
    }

    public String getProduct() {
        return product;
    }

    public void setProduct(String product) {
        this.product = product;
    }

    public BigDecimal getPrice() {
        return price;
    }

    public void setPrice(BigDecimal price) {
        this.price = price;
    }
}