# implementation/at.electro.shop.bff/src/main/java/at/electro/shop/bff/api/notification/models/ProductMarkingResponse.java
package at.electro.shop.bff.api.notification.models;

public class ProductMarkingResponse extends ProductMarkingRequest {
    private String uuid;

    public String getUuid() {
        return uuid;
    }

    public void setUuid(String uuid) {
        this.uuid = uuid;
    }

}