# implementation/at.electro.shop.bff/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/vendor/{vendorId}/inventory":{"get":{"tags":["inventory-controller"],"operationId":"getInventory","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventoryResponse"}}}}}},"put":{"tags":["inventory-controller"],"operationId":"putInventory","parameters":[{"name":"vendorId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventoryResponse"}}}}}}},"/api/v1/products/{uuid}/rating":{"put":{"tags":["product-controller"],"operationId":"updateProductRating","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/products/{productId}":{"get":{"tags":["product-controller"],"operationId":"getProduct_1","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["product-controller"],"operationId":"putProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/products/{productId}/price":{"put":{"tags":["product-controller"],"operationId":"updatePrice","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/securedPage":{"post":{"tags":["auth-controller"],"operationId":"securedPage","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/transactions":{"get":{"tags":["transaction-controller"],"operationId":"transactions","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}},"post":{"tags":["transaction-controller"],"operationId":"post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}}},"/api/v1/shipment/{uuid}/status/{status}":{"post":{"tags":["shipment-controller"],"operationId":"postShipmentUpdateOrderConfirmation","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"path","required":true,"schema":{"type":"string","enum":["OrderConfirmation","OutForDelivery","Delivered"]}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/products":{"get":{"tags":["product-controller"],"operationId":"getProduct","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"post":{"tags":["product-controller"],"operationId":"postProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/api/v1/notification/productmarking":{"post":{"tags":["notification-controller"],"operationId":"addProductMarking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductMarkingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductMarkingResponse"}}}}}}},"/api/v1/api/v1/users/{userId}/cart/product/{productId}":{"post":{"tags":["user-controller"],"operationId":"saveProduct","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}},"delete":{"tags":["user-controller"],"operationId":"removeProduct","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"productId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}}},"/api/v1/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"/api/v1/users/{userId}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/users/{userId}/transactions":{"get":{"tags":["transaction-controller"],"operationId":"getUserTransactions","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}}},"/api/v1/users/{userId}/cart":{"get":{"tags":["user-controller"],"operationId":"getCart","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}}},"/api/v1/user/me":{"get":{"tags":["user-controller"],"operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/shipments":{"get":{"tags":["shipment-controller"],"operationId":"get","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShipmentUpdate"}}}}}}}},"/api/v1/sellers/{sellerId}/transactions":{"get":{"tags":["transaction-controller"],"operationId":"getSellerTransactions","parameters":[{"name":"sellerId","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}}},"/api/v1/notification/productmarking/user/{userid}":{"get":{"tags":["notification-controller"],"operationId":"getMarkings","parameters":[{"name":"userid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductMarkingResponse"}}}}}}}},"/api/v1/buyers/{buyerId}/transactions":{"get":{"tags":["transaction-controller"],"operationId":"getBuyerTransactions","parameters":[{"name":"buyerId","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"createdAt"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}}}}}},"/api/v1/api/v1/sellers/{sellerId}/profits":{"get":{"tags":["transaction-controller"],"operationId":"getProfits","parameters":[{"name":"sellerId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number"}}}}}}},"/":{"get":{"tags":["auth-controller"],"operationId":"index","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/notification/productmarking/{id}":{"delete":{"tags":["notification-controller"],"operationId":"deleteMarking","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductMarkingResponse"}}}}}}}},"components":{"schemas":{"InventoryRequest":{"type":"object","properties":{"inventoryUuid":{"type":"string"},"vendor":{"type":"string"},"productIdListToRemove":{"type":"array","items":{"type":"string"}},"productsToAdd":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}},"Product":{"type":"object","properties":{"uuid":{"type":"string"},"name":{"type":"string"},"brand":{"type":"string"},"img":{"type":"string"},"rating":{"type":"number"},"ratingsCount":{"type":"integer","format":"int32"},"description":{"type":"string"},"alcoholPercent":{"type":"number"},"price":{"type":"number"},"currency":{"type":"string"}}},"InventoryResponse":{"type":"object","properties":{"inventoryUuid":{"type":"string"},"vendor":{"type":"string"},"inventoryValue":{"type":"number","format":"double"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductResult"}}}},"ProductResult":{"type":"object","properties":{"amount":{"type":"integer","format":"int32"},"product":{"$ref":"#/components/schemas/Product"}}},"ProductRequest":{"type":"object","properties":{"vendor":{"type":"string"},"name":{"type":"string"},"brand":{"type":"string"},"img":{"type":"string"},"rating":{"type":"number"},"ratingsCount":{"type":"integer","format":"int32"},"description":{"type":"string"},"alcoholPercent":{"type":"number"},"price":{"type":"number"},"currency":{"type":"string"}}},"ProductResponse":{"type":"object","properties":{"uuid":{"type":"string"},"name":{"type":"string"},"brand":{"type":"string"},"img":{"type":"string"},"rating":{"type":"number"},"ratingsCount":{"type":"integer","format":"int32"},"description":{"type":"string"},"alcoholPercent":{"type":"number"},"price":{"type":"number"},"currency":{"type":"string"}}},"Address":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"street":{"type":"string"},"streetNumber":{"type":"string"},"doorNumber":{"type":"string"},"floor":{"type":"string"},"city":{"type":"string"},"zipCode":{"type":"string"},"country":{"type":"string"}}},"TransactionProduct":{"type":"object","properties":{"id":{"type":"string"},"price":{"type":"number"}}},"TransactionRequest":{"type":"object","properties":{"seller":{"type":"string"},"buyer":{"type":"string"},"currency":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"modifiedAt":{"type":"string","format":"date-time"},"address":{"$ref":"#/components/schemas/Address"},"paymentMethod":{"type":"string","enum":["PAYMENT_ON_DELIVERY"]},"products":{"type":"array","items":{"$ref":"#/components/schemas/TransactionProduct"}}}},"TransactionResponse":{"type":"object","properties":{"uuid":{"type":"string"},"seller":{"type":"string"},"buyer":{"type":"string"},"product":{"type":"string"},"price":{"type":"number"},"currency":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"modifiedAt":{"type":"string","format":"date-time"},"paymentMethod":{"type":"string","enum":["PAYMENT_ON_DELIVERY"]},"address":{"$ref":"#/components/schemas/Address"}}},"ProductMarkingRequest":{"type":"object","properties":{"user":{"type":"string"},"emailAddress":{"type":"string"},"product":{"type":"string"},"price":{"type":"number"}}},"ProductMarkingResponse":{"type":"object","properties":{"user":{"type":"string"},"emailAddress":{"type":"string"},"product":{"type":"string"},"price":{"type":"number"},"uuid":{"type":"string"}}},"CartDto":{"type":"object","properties":{"uuid":{"type":"string"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}},"ProductDto":{"type":"object","properties":{"count":{"type":"integer","format":"int32"},"id":{"type":"string"}}},"AddressDto":{"type":"object","properties":{"street":{"type":"string"},"streetNumber":{"type":"string"},"doorNumber":{"type":"integer","format":"int32"},"floor":{"type":"string"},"city":{"type":"string"},"zipCode":{"type":"integer","format":"int32"},"country":{"type":"string"}}},"UserDto":{"type":"object","properties":{"uuid":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"},"cart":{"$ref":"#/components/schemas/CartDto"}}},"ShipmentUpdate":{"type":"object","properties":{"uuid":{"type":"string"},"address":{"$ref":"#/components/schemas/Address"},"product":{"type":"string"},"status":{"type":"string","enum":["OrderConfirmation","OutForDelivery","Delivered"]}}}}}}